Make getProviderMimeTypeAsync really asynchronous
Oneway binders calls to the same binder object will be dispatched
one by one, that means if one of the call are blocked, the rest
will all be blocked too, see the binder driver for more detail.

When someone calls ContentProvider#getType(), and if the target
provider was not already acuqired by the calling process, the
actuall call will be delegagted into system_server, because this
API was designed to be available for everyone, with or without
certain permissions.

Even AMS#getProviderMimeTypeAsync is declared as an oneway binder
call, but it still needs to wait for the remote ContentProvider to
be correctly published, this wait might take 10~20 seconds in some
scenarios. In the meantime, if any other oneway binder calls were
invoked on AMS, they shall be blocked until the provider is ready
and if any of the blocked oneway binder calls happens to be
AMS#serviceDoneExecuting() and the like, the user will see an
unexpected ANR dialogs which totaly doesn't make any sense since
these apps are certainly not unresponsive.

Bug: 122015758
Test: atest FrameworksCoreTests:android.content.ContentResolverTest

Signed-off-by: wangmingming1 <wangmingming1@xiaomi.com>
Change-Id: I74b8fa10233b9a475c252b3abad638d43f040994